begin to use database, refs #142.
